Replace RunWithTransaction with ExtractTransaction
RunWithTransaction runs a callback with a transaction found in the
context. This is somewhat an awkward interface with the current uses.

Sometimes we want to actually extract the transaction from the context,
not just do things with it. Doing so currently requires declaring a
variable outside of the callback and assigning into it within the
callback.

Sometimes we need to return an error from the callback. As we don't
support propagating an error, the error variable needs to be declared
outside of the callback, assigned in the callback, and then checked
outside of the callback.

This commit replaces the function with ExtractTranscation that simply
extracts the transaction from the context. This is still convenient for
uses when we just need to call something on the transaction but is a lot
more convenient for the above two cases. For the first case, we can now
simply extract the transaction into the target variable. For the second
case, we can simply return directly from the if block if an error occurs.

+++ internal/gitaly/service/objectpool/delete.go
@@ -5,7 +5,6 @@ import (
 	"errors"
 
 	"gitlab.com/gitlab-org/gitaly/v16/internal/git/objectpool"
-	"gitlab.com/gitlab-org/gitaly/v16/internal/gitaly/storage"
 	"gitlab.com/gitlab-org/gitaly/v16/internal/gitaly/storage/storagectx"
 	"gitlab.com/gitlab-org/gitaly/v16/internal/structerr"
 	"gitlab.com/gitlab-org/gitaly/v16/proto/go/gitalypb"
@@ -27,9 +26,9 @@ func (s *server) DeleteObjectPool(ctx context.Context, in *gitalypb.DeleteObject
 		return nil, structerr.NewInternal("%w", err)
 	}
 
-	storagectx.RunWithTransaction(ctx, func(tx storage.Transaction) {
+	if tx := storagectx.ExtractTransaction(ctx); tx != nil {
 		tx.DeleteRepository()
-	})
+	}
 
 	return &gitalypb.DeleteObjectPoolResponse{}, nil
 }
